Struct blender_mesh::BlenderMesh

pub struct BlenderMesh { /* fields omitted */ }

All of the data about a mesh

TODO: Rename crate to MeshIr

Implementations

impl BlenderMesh

pub fn combine_vertex_indices(
&mut self,
config: &CreateSingleIndexConfig
) -> SingleIndexedVertexAttributes

We store our exported Blender mesh with indices for positions, normals and uvs because

Easier because we we can unit test that here vs. a blender python script that’s much trickier to test.

Smaller to store the fields indexed than to expand them.

Most rendering pipelines only supports one index buffer, so here we convert our vertex data from having three indices to having one.

This typically requires some duplication of vertex data - we duplicate the minimum amount of vertex data necessary.

TODO: Need to continue refactoring

TODO: Make this function set BlenderMesh.vertex_attributes = VertexData::SingleIndexVertexData

TODO: Don’t work on additionally functionality until we’ve broken up these tests and implementation into smaller, specific pieces.

TODO: There are unexpected (based on the method’s name) mutations in here such as triangulation. Lot’s to refactor in this crate.

impl BlenderMesh

pub fn interleave_vertex_data(&self) -> Vec<u8>

impl BlenderMesh

pub fn y_up(&mut self)

Blender meshes get exported with a Z up coordinate system. Here we flip our coordinate system to be y up

@see https://gamedev.stackexchange.com/a/7932

TODO: When we have bone data we’ll need to change them to port change-mat4-coordinate-system into here. https://github.com/chinedufn/change-mat4-coordinate-system/blob/master/change-mat4-coordinate-system.js

impl BlenderMesh

pub fn pbr_cube_without_textures() -> Self

Create a default Blender cube with uniform PBR texture inputs.

A 2x2x2 cube centered about the origin.

impl BlenderMesh

pub fn armature_name(&self) -> Option<&String>

The name of this mesh’s parent armature

pub fn set_armature_name(&mut self, armature_name: Option<String>)

Set the name of this mesh’s parent armature

pub fn materials(&self) -> &HashMap<String, PrincipledBSDF>

A map of material name to the material’s data

pub fn materials_mut(&mut self) -> &mut HashMap<String, PrincipledBSDF>

A mutable map of material name to the material’s data

pub fn custom_properties(&self) -> &HashMap<String, CustomProperty>

Custom properties for this mesh

i.e. in Blender this might be found with bpy.context.view_layer.objects.active.keys()

pub fn bounding_box(&self) -> BoundingBox

The smallest box that contains the entire mesh

pub fn set_bounding_box(&mut self, bounding_box: BoundingBox)

Set the mesh’s bounding box.

pub fn name(&self) -> &String

The name of the mesh

pub fn set_name(&mut self, name: String)

Set the name of the mesh